+++ tests/unit/test_expression_tree/test_unary_operators.py
@@ -51,6 +51,34 @@ class TestUnaryOperators(unittest.TestCase):
             np.diag(signb.evaluate().toarray()), [-1, -1, 0, 1, 1]
         )
 
+    def test_floor(self):
+        a = pybamm.Symbol("a")
+        floora = pybamm.Floor(a)
+        self.assertEqual(floora.name, "floor")
+        self.assertEqual(floora.children[0].name, a.name)
+
+        b = pybamm.Scalar(3.5)
+        floorb = pybamm.Floor(b)
+        self.assertEqual(floorb.evaluate(), 3)
+
+        c = pybamm.Scalar(-3.2)
+        floorc = pybamm.Floor(c)
+        self.assertEqual(floorc.evaluate(), -4)
+
+    def test_ceiling(self):
+        a = pybamm.Symbol("a")
+        ceila = pybamm.Ceiling(a)
+        self.assertEqual(ceila.name, "ceil")
+        self.assertEqual(ceila.children[0].name, a.name)
+
+        b = pybamm.Scalar(3.5)
+        ceilb = pybamm.Ceiling(b)
+        self.assertEqual(ceilb.evaluate(), 4)
+
+        c = pybamm.Scalar(-3.2)
+        ceilc = pybamm.Ceiling(c)
+        self.assertEqual(ceilc.evaluate(), -3)
+
     def test_gradient(self):
         # gradient of scalar symbol should fail
         a = pybamm.Symbol("a")
@@ -255,6 +283,12 @@ class TestUnaryOperators(unittest.TestCase):
 
         # sign
         self.assertEqual((pybamm.sign(a)).diff(a).evaluate(y=y), 0)
+
+        # floor
+        self.assertEqual((pybamm.Floor(a)).diff(a).evaluate(y=y), 0)
+
+        # ceil
+        self.assertEqual((pybamm.Ceiling(a)).diff(a).evaluate(y=y), 0)
 
         # spatial operator (not implemented)
         spatial_a = pybamm.SpatialOperator("name", a)